Version 1.0
In the version 1.0, the customer invoice printout is modified to adapt the requirements from Sarinah Motor. It creates two paper formats (A4 and A5) with custom margins. The header and footer have been changed as well. In the document, tax related columns and fields are now hidden. Now basically it is invoice without payment information.