Add Giro fields and make Giro available for reporting
Add giro_number and giro_create_date for Payment by Giro purpose. This update also add journal_name field that is used to make invisible the Giro related field if the payment method chosen is not Giro.
The XML view used includes Widget as well as Form because the widget is used when choosing payment method.